Load computers, ips and interfaces from AWS Config
RDS is a bit wierd in that is hard to work out what ENI it is using.
Most other kinds of computers ENIs map to will have an id somewhere
on the ENI config, RDS does not. We have to symlink all possible ips
for evry ENI and then lookup the IP of the endpoint domain name we get
when loading the RDS config.